Add a network to a multiplex, a multilayer or an hummus object

Add a network to a multiplex, a multilayer or an hummus object

Usage

add_network(
  object,
  network,
  network_name,
  multiplex_name = NULL,
  directed = FALSE,
  weighted = FALSE,
  verbose = 1
)

Arguments

object

A multiplex, a multilayer or an hummus object

network

A network (edge list)

network_name

The name of the network

multiplex_name

The name of the multiplex. Default: NULL if object is a multiplex already only

directed

Logical indicating if the network is directed. Default: FALSE

weighted

Logical indicating if the network is weighted. Default: FALSE

verbose

(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.

Value

A multiplex, a multilayer or an hummus object with the added network

Examples

hummus <- add_network(
                           object = hummus,
                           network = atac_peak_network,
                           network_name = network_name,
                           multiplex_name = multiplex_name,
                           weighted = TRUE,
                           directed = FALSE)
#> Error: object 'hummus' not found