Compute links between DNA regions and genenames

Compute and add bipartite between DNA regions and genenames to hummus object. Links are computed based on the distance between peaks and gene's TSS location from gene.range annotations. Call find_peaks_near_genes function, that can use different methods.

Usage

bipartite_peaks2genes(
  hummus_object,
  gene_assay = "RNA",
  peak_assay = "peaks",
  gene_multiplex_name = NULL,
  peak_multiplex_name = NULL,
  peak_to_gene_method = "Signac",
  upstream = 500,
  downstream = 500,
  only_tss = TRUE,
  store_network = FALSE,
  output_file = NULL,
  bipartite_name = "atac_rna"
)

Arguments

hummus_object

(hummus_object) - Hummus object.

gene_assay

(character) - Name of assay containing the gene expression data. Default: "RNA".

peak_assay

(character) - Name of the assay containing the DNA regions (ATAC peaks). Default: "peaks".

gene_multiplex_name

(character) - Name of the multiplex containing the genes. If NULL, the name of the gene assay is used.

peak_multiplex_name

(character) - Name of the multiplex containing the DNA regions (ATAC peaks). If NULL, the name of the peak assay is used.

peak_to_gene_method

(character) - Method to use to compute the links between peaks and genes. Default: "Signac".

'Signac' - Use Signac::Extend to extend genes.

'GREAT' - Not implemented yet.

upstream

(int) - Upstream distance from TSS to consider as potential promoter.

downstream

(int) - Downstream distance from TSS to consider as potential promoter.

only_tss

(logical) - If TRUE, only TSS will be considered.

store_network

(bool) - Save the bipartite directly (TRUE, default) or return without saving on disk (FALSE).

output_file

(character) - Name of the output_file (if store_bipartite == TRUE). Default: NULL.

bipartite_name

(character) - Name of bipartite. Default: "atac_rna".

verbose

(integer) Display function messages. Set to 0 for no message displayed, >= 1 for more details. Default: 1.

Value

hummus_object (hummus_object) - Hummus object w/ atac-rna bipartite added to the multilayer slot

Examples

hummus <- bipartite_peaks2genes(
                       hummus_object = hummus,
                       gene_assay = "RNA",
                       peak_assay = "peaks",
                       gene_multiplex_name = "RNA",
                       peak_multiplex_name = "peaks",
                       peak_to_gene_method = "Signac",
                       upstream = 500,
                       downstream = 500,
                       only_tss = TRUE,
                       store_network = FALSE,
                       bipartite_name = "atac_rna")
#> Error: object 'hummus' not found