Compute peak network from scATAC-seq data

This function will create a network from atac data (or in theory any data wtih peaks coordinates as features). Different method should be implemented at some point (e.g. RENIN), for now Cicero is still the reference and only method available

Usage

compute_atac_peak_network(
  hummus,
  atac_assay = "peaks",
  genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
  method = "cicero",
  multiplex_name = NULL,
  network_name = NULL,
  store_network = FALSE,
  output_file = NULL,
  threshold = 0,
  number_cells_per_clusters = 50,
  sample_num = 100,
  seed = 2025,
  verbose = 1,
  window = 5e+05,
  reduction_method = "UMAP"
)

Arguments

hummus

(Hummus_Object) - Hummus object

atac_assay

(character) - Name of the assay containing the atac peaks data.

genome

(BSgenome) - Genome used to compute the distance between peaks.

method

(character) - Method used to infer network edges.

'cicero' - Use cicero to infer regulatory networks.

'Other method' - TO DO.

multiplex_name

(character) - Name of the multiplex to add the network to. Default is 'peaks'.

network_name

(character) - Name of the network in the multiplex to add the network to. Default is 'peak_network'.

store_network

(bool) - Save the network directly (TRUE, default) or return without saving on disk (FALSE).

output_file

(character) - Name of the output_file (if store_network == TRUE).

threshold

(interger, default 0) - Minimal threshold to select tf-gene edges.

number_cells_per_clusters

(integer) - Number of cells grouped by territory to define pseudocells

sample_num

(integer | Cicero) - Number of pseudocells to sample from each territory. Default is 100.

seed

(integer | Cicero) - Seed used to sample pseudocells. Default is 2025

verbose

(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.

window

(integer) - Size of window to consider potential cis-regulatory cooperations between peaks. Default is 500K base pairs.

reduction_method

(character | Cicero) - Method used to reduce dimensionality of the data to identify territories. Default is 'UMAP'.

Value

(data.frame) - Return list of network interactions between peaks

Details

Method descriptions :

Cicero Use patial corelation between peaks that are in a given window (e.g. : less distant than 500K base pairs)

Examples

hummus <- compute_atac_peak_network(hummus)
#> Error: object 'hummus' not found