Compute gene netwok from scRNA-seq data

This function will create a network from rna data (or in theory any data wtih genes as features). Different method should be implemented at some point (any suggestion is welcomed ! :) ), for now Genie3 is still the reference and only method available

compute_gene_network(
  hummus,
  gene_assay = "RNA",
  tfs = NULL,
  method = "GENIE3",
  multiplex_name = NULL,
  network_name = NULL,
  store_network = FALSE,
  output_file = NULL,
  threshold = 0,
  number_cores = 1,
  verbose = 1
)

Arguments

hummus

(hummus_object) - Hummus object

gene_assay

(character) - Name of the assay containing the gene expression data.

tfs

vector(character) - List of tfs considered. If NULL, all TFs with motifs in the hummus object are used.

method

(character) - Method used to infer network edges.

'Genie3' - Use tree random forest to infer regulatory networks.

'Other method' - TO DO.

multiplex_name

(character) - Name of the multiplex to add the network to. Default is 'RNA'.

network_name

(character) - Name of the network in the multiplex to add the network to. Default is 'RNA_network'.

store_network

(bool) - Save the network directly (TRUE, default) or return without saving on disk (FALSE).

output_file

(character) - Name of the output_file (if store_network == TRUE).

threshold

(interger, default 0) - Minimal threshold to select tf-gene edges.

number_cores

(interger, default 1) - Number of thread that should be used for the parallelizable methods.

verbose

(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.

Value

(data.frame) - Return list of network interactions between genes

Details

Method descriptions :

Genie3 Use tree random forest to infer regulatory networks : https://bioconductor.org/packages/release/bioc/html/GENIE3.html

Examples

hummus <- compute_gene_network(
                               hummus,
                               gene_assay = "RNA",
                               method = "GENIE3",
                               verbose = 1,
                               number_cores = 8,
                               store_network = FALSE)
#> Computing gene network with  GENIE3  ...
#> 	No TFs list provided, fetching from hummus object...
#> Error in eval(expr, envir, enclos): object 'hummus' not found