Compute TF network and add it to hummus object

Compute a protein-protein interaction layer from Omnipath request that will represent tf cooperativity. This network is the top-layer of HuMMuS multilayer.

compute_tf_network(
  hummus = null,
  organism = 9606,
  tfs = NA,
  gene_assay = NULL,
  store_network = FALSE,
  output_file = NULL,
  source_target = "AND",
  multiplex_name = "TF",
  tf_network_name = "TF_network",
  verbose = 1
)

Arguments

hummus

(hummus_object) - Hummus object

organism

(integer) - Specie identifier from Omnipath to fetch specific interactions

tfs

vector(character) - List of tfs considered.

gene_assay

(character) - Name of the assay to get tfs from if tfs is not provided. If NULL, all TFs with motifs in the hummus object are used.

store_network

(bool) - Save the network directly (TRUE, default) or return without saving on disk (FALSE).

output_file

(character) - Name of the output_file (if store_network == TRUE).

source_target

('AND'|'OR') - Fetch only the interactions involving two considered tfs ('AND', default), or one considered tfs and any other element ('OR')

multiplex_name

(character) - Name of the multiplex to add the network to. Default is 'TF'.

tf_network_name

(character) - Name of the network in the multiplex to add the network to. Default is 'TF_network'.

verbose

(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.

Value

(hummus_object) - Return hummus object with the new network added.

Examples

hummus <- compute_tf_network(hummus,
                                       gene_assay = "RNA",
                                       verbose = 1)
#> Computing TF network...
#> Error in eval(expr, envir, enclos): object 'hummus' not found