Define enhancers from hummus object

Calling the define_output function with output_type = 'enhancers'

Usage

define_enhancers(
  hummus_object,
  multiplex_names = NULL,
  bipartites_names = NULL,
  config_name = "enhancers_config.yml",
  config_folder = "config",
  tf_multiplex = "TF",
  atac_multiplex = "peaks",
  rna_multiplex = "RNA",
  multilayer_f = "multilayer",
  gene_list = NULL,
  tf_list = NULL,
  save = FALSE,
  output_f = NULL,
  return_df = TRUE,
  suffix_bipartites = ".tsv",
  njobs = 1
)

Arguments

hummus_object

A hummus object

multiplex_names

A vector of multiplex names considered. It must be a subset of the names of the multiplexes in the hummus object.

bipartites_names

A vector of bipartites names considered. It must be a subset of the names of the bipartites in the hummus object.

config_name

The name of the config file to be created by hummuspy

config_folder

The folder where the config file will be created

tf_multiplex

The name of the multiplex containing the TFs

atac_multiplex

The name of the multiplex containing the ATAC-seq peaks

rna_multiplex

The name of the multiplex containing the RNA-seq genes

multilayer_f

The folder where the multilayer is stored

gene_list

A vector of genes to be considered for the final enhancers (filtering is done on the genes before inferring the enhancers)

tf_list

A vector of TFs to be considered for the final enhancers (filtering is done on the TFs after inferring the enhancers)

save

A boolean indicating if the enhancers should be saved

output_f

The name of the file where the enhancers should be saved (if save == TRUE)

return_df

A boolean indicating if the enhancers should be returned as a dataframe

suffix_bipartites

A suffix to add to the bipartites names (to indicate the exact file location)

njobs

The number of jobs to be used for to compute of the enhancers

Value

A dataframe containing the enhancers (if return_df == TRUE)

Examples

enhancers <- define_enhancers(hummus_object = hummus,
                            multilayer_f = multilayer_folder,
                            njobs = 5)
#> Error in eval(expr, envir, enclos): object 'hummus' not found