Define general config file for hummuspy

Define general config file for hummuspy

Usage

define_general_config(
  hummus_object,
  multiplex_names = NULL,
  bipartites_names = NULL,
  folder_multiplexes = "multiplex",
  folder_bipartites = "bipartites",
  seed_path = "seed/seeds.txt",
  suffix = ".tsv",
  self_loops = FALSE,
  restart_proba = 0.7,
  save_configfile = FALSE,
  config_name = "config.yml",
  config_folder = "config",
  multilayer_f = "multilayer",
  suffix_bipartites = ".tsv"
)

Arguments

hummus_object

A hummus object

multiplex_names

A vector of multiplex names considered. It must be a subset of the names of the multiplexes in the hummus object, or NULL if all multiplexes should be considered.

bipartites_names

A vector of bipartites names considered. It must be a subset of the names of the bipartites in the hummus object, or NULL if all bipartites should be considered.

folder_multiplexes

The folder where the multiplexes are stored

folder_bipartites

The folder where the bipartites are stored

seed_path

The path to the seed file

self_loops

A boolean indicating if self loops should be considered.

restart_proba

The restart probability for the random walk (default = 0.7)

save_configfile

A boolean indicating if the config file should be saved

config_name

The name of the config file to be created by hummuspy

config_folder

The folder where the config file will be created (inside multilayer_f)

multilayer_f

The folder where the multilayer is stored

suffix_bipartites

A suffix to add to the bipartites names (to indicate the exact file name)

Value

A config file for hummuspy