Define output from hummus object

Define output from hummus object

define_output(
  output_type,
  hummus_object,
  multiplex_names = NULL,
  bipartites_names = NULL,
  config_name = "target_genes_config.yml",
  config_folder = "config",
  tf_multiplex = "TF",
  atac_multiplex = "peaks",
  rna_multiplex = "RNA",
  multilayer_f = "multilayer",
  gene_list = NULL,
  tf_list = NULL,
  save = FALSE,
  output_f = NULL,
  return_df = TRUE,
  suffix_bipartites = ".tsv",
  njobs = 1
)

Arguments

output_type

The type of output to be defined

hummus_object

A hummus object

multiplex_names

A vector of multiplex names considered. It must be a subset of the names of the multiplexes in the hummus object.

bipartites_names

A vector of bipartites names considered. It must be a subset of the names of the bipartites in the hummus object.

config_name

The name of the config file to be created by hummuspy

config_folder

The folder where the config file will be created

tf_multiplex

The name of the multiplex containing the TFs

atac_multiplex

The name of the multiplex containing the ATAC-seq peaks

rna_multiplex

The name of the multiplex containing the RNA-seq genes

multilayer_f

The folder where the multilayer is stored

gene_list

A vector of genes to be considered for the target_genes (filtering is done on the genes before inferring the target_genes)

tf_list

A vector of TFs to be considered for the final target_genes (filtering is done on the TFs after inferring the target_genes)

save

A boolean indicating if the target_genes should be saved

output_f

The name of the file where the target_genes should be saved (if save == TRUE)

return_df

A boolean indicating if the target_genes should be returned as a dataframe

suffix_bipartites

A suffix to add to the bipartites names (to indicate the exact file location)

njobs

The number of jobs to be used to compute of the target_genes

Value

A dataframe containing the target_genes (if return_df == TRUE)

Examples

target_genes <- define_output('grn', hummus_object = hummus)
#> Error in eval(expr, envir, enclos): object 'hummus' not found