Format bipartites names for python hummuspy package config functions

Format bipartites names for python hummuspy package config functions

Usage

format_bipartites_names(
  hummus_object,
  bipartites_names = NULL,
  suffix_bipartites = ".tsv"
)

Arguments

hummus_object

A hummus object

bipartites_names

A vector of bipartites names considered. It must be a subset of the names of the bipartites in the hummus object.

suffix_bipartites

A suffix to add to the bipartites location

Value

A list of bipartites names formatted for hummuspy config funtions each element of the list is a list containing the right and left layer connected by the bipartite

Examples

bipartites_dictionary <- format_bipartites_names(
                                      hummus_object = hummus,
                                      bipartites_names = c("atac_rna",
                                                           "tf_peaks"))
#> Error: object 'hummus' not found