Cicero wrapper function for the compute_atac_peak_network function

This function is a wrapper for the compute_atac_peak_network function in layers.R. It computes the peak network from scATAC-seq data using Cicero. It returns a data frame with the peak network. The data frame also contains the coaccess score for each edge. The coaccess score is the probability that two peaks are accessible in the same cell. The coaccess score is computed by Cicero. Edges are filtered based on the coaccess score. Only edges with a coaccess score > threshold are kept.

run_cicero_wrapper(
  hummus,
  atac_assay,
  genome,
  window,
  number_cells_per_clusters,
  sample_num,
  seed,
  verbose,
  threshold,
  reduction_method = "UMAP"
)

Arguments

hummus

A hummus object

atac_assay

The name of the assay containing the scATAC-seq data

genome

The genome object

window

The window size used by Cicero to compute the coaccess score

number_cells_per_clusters

The number of cells per cluster used by Cicero to compute the coaccess score

sample_num

The number of samples used by Cicero to compute the coaccess score

seed

The seed used by Cicero to compute the coaccess score

verbose

The verbosity level

threshold

The threshold used to filter edges based on the coaccess score

reduction_method

The method used by monocle3 to reduce the dimension of the scATAC-seq data before defining the pseudocells. The default is UMAP.

Value

A data frame containing the peak network