Save multilayer object files in a hierarchical structure on disk

Save multilayer files from a Hummus_Object in a hierarchical structure on disk, inside a folder specified through folder_name

Usage

save_multilayer(hummus, folder_name, verbose = TRUE, suffix = ".tsv")

Arguments

hummus

A hummus object

folder_name

The name of the folder to save the multilayer

verbose

(integer) - Display function messages. Set to 0 for no message displayed, >= 1 for more details.

suffix

The suffix of the files to save. Default: ".tsv"

Value

Nothing, but create a folder containing the multilayer object files

Examples

folder_name = "multilayer"
save_multilayer(hummus = hummus, folder_name = "multilayer")
#> Error: object 'hummus' not found