Compute links between TFs and DNA regions (ATAC peaks)

Compute and add bipartite between TFs and DNA regions to hummus object. Links are computed based on the binding motifs of TFs and their locations on a reference genome. Currently based on Signac AddMotifs function (–> motifmachR, itself based on MOODs algorithm).

Usage

bipartite_tfs2peaks(
  hummus_object,
  tf_expr_assay = "RNA",
  peak_assay = "peaks",
  tf_multiplex_name = NULL,
  peak_multiplex_name = NULL,
  genome,
  store_network = FALSE,
  output_file = NULL,
  verbose = 1,
  bipartite_name = "tf_peak"
)

Arguments

hummus_object

(hummus_object) - Hummus object.

tf_expr_assay

(character) - Name of assay containing the TF expression data. If NULL, all TFs with a motif are used. Default: "RNA".

peak_assay

(character) - Name of the assay containing the DNA regions (ATAC peaks). Default: "peaks".

tf_multiplex_name

(character) - Name of multiplex containing the TFs. If NULL, the name of the TF assay is used.

peak_multiplex_name

(character) - Name of the multiplex containing the DNA regions (ATAC peaks). If NULL, the name of the peak assay is used.

genome

(BSgenome object) - Reference genome.

store_network

(bool) - Save the bipartite directly (TRUE, default) or return without saving on disk (FALSE).

output_file

(character) - Name of the output_file (if store_bipartite == TRUE). Default: NULL.

verbose

(integer) Display function messages. Set to 0 for no message displayed, >= 1 for more details. Default: 1.

bipartite_name

(character) - Name of bipartite. Default: "tf_peak".

Value

hummus_object (hummus_object) - Hummus object with TF-peak bipartite added to the multilayer slot

Examples

hummus <- bipartite_tfs2peaks(
                     hummus_object = hummus,
                     tf_expr_assay = "RNA",
                     peak_assay = "peaks",
                     tf_multiplex_name = "TF",
                     peak_multiplex_name = "peaks",
          genome = BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38,
                     store_network = FALSE,
                     verbose = 1,
                     bipartite_name = "tf_peak")
#> Computing TF-peak bipartite
#> Error: object 'hummus' not found